// src/pages/types.ts
// export interface Post {
//   id: string;
//   companyCode: string;
//   listingStatus: string;
//   companyName: string;
//   companyNameRead: string;
//   postCode: string;
//   location: string;
//   representative: string;
//   representativeRead: string;
//   phone: string;
//   Earnings2022: string;
//   profit2022: string;
//   Earnings2021: string;
//   profit2021: string;
//   Earnings2020: string;
//   profit2020: string;
//   created_at: string;
//   updated_at: string;
// }

export interface Post {
  id: string;
  companyCode: number;
  listingStatus: string;
  companyName: string;
  companyNameRead: string;
  postCode: string;
  location: string;
  representative: string;
  representativeRead: string;
  phone: string;
  Earnings2022: string;
  profit2022: string;
  Earnings2021: string;
  profit2021: string;
  Earnings2020: string;
  profit2020: string;
  created_at: string;
  updated_at: string;
}